eCommerce Yubit - Services Controller

+++ api/src/main/java/com/yubit/api/core/product/Product.java
@@ -1,4 +1,4 @@
 package com.yubit.api.core.product;
 
-public record Product(Long id,String name, int weight, String serviceAddress) {
+public record Product(int id,String name, int weight, String serviceAddress) {
 }
