+++ api/src/main/java/com/yubit/api/core/product/ProductService.java
@@ -8,5 +8,5 @@ public interface ProductService {
             value = "/product/{productId}",
             produces = "application/json"
     )
-    Product getProduct(@PathVariable Long productId);
+    Product getProduct(@PathVariable int productId);
 }
